Add support for sibling nodes
Sibling nodes are nodes that are nearer of other nodes than from the s3 bucket.

This usually happens with third party VPS services.

Eg

```
  "siblings": {
    "cdn-southafrica-1": "cdn-eu-central"
  }
```

This means that the node in South Africa uses Europe Central as a backend instead of S3.

If EU Central is down in the above example both nodes will be down.